Renamed MediaTracking class to ImageReferenceFactory
