+++ src/Skybrud.ImagePicker/Composers/ImagePickerComposer.cs
@@ -1,4 +1,4 @@
-﻿using Skybrud.ImagePicker.Tracking;
+﻿using Skybrud.ImagePicker.Factories;
 using Umbraco.Core;
 using Umbraco.Core.Composing;
 
@@ -7,7 +7,7 @@ namespace Skybrud.ImagePicker.Composers {
     public class ImagePickerComposer : IUserComposer {
         
         public void Compose(Composition composition) {
-            composition.DataValueReferenceFactories().Append<MediaTracking>();
+            composition.DataValueReferenceFactories().Append<ImageReferenceFactory>();
         }
 
     }
